takeoff() method in Flight
